Add install scripts for Windows 10
Why does this have more in it than the Linux scripts? Especially given the fact that I almost never use windows? It's simple, I want to deal with Windows as little as possible when I do have the misfortune of encountering it.

This covers pretty much anything I could even come up with needing on a Windows install, some stuff may be missing though. <3 gsudo btw